8211977: move testlibrary tests into one place
