+++ make/Main.gmk
@@ -640,6 +640,18 @@ $(eval $(call SetupTarget, test-image-jdk-jtreg-native, \
     DEPS := build-test-jdk-jtreg-native, \
 ))
 
+$(eval $(call SetupTarget, build-test-libtest-jtreg-native, \
+    MAKEFILE := test/JtregNativeLibTest, \
+    TARGET := build-test-libtest-jtreg-native, \
+    DEPS := buildtools-jdk, \
+))
+
+$(eval $(call SetupTarget, test-image-libtest-jtreg-native, \
+    MAKEFILE := test/JtregNativeLibTest, \
+    TARGET := test-image-libtest-jtreg-native, \
+    DEPS := build-test-libtest-jtreg-native, \
+))
+
 $(eval $(call SetupTarget, build-test-hotspot-jtreg-graal, \
     MAKEFILE := test/JtregGraalUnit, \
     TARGET := build-test-hotspot-jtreg-graal, \
@@ -982,6 +994,7 @@ else
   interim-image: $(INTERIM_JMOD_TARGETS)
 
   build-test-hotspot-jtreg-native: hotspot-$(JVM_VARIANT_MAIN)-libs
+  build-test-libtest-jtreg-native: hotspot-$(JVM_VARIANT_MAIN)-libs
 
 endif
 
@@ -1105,7 +1118,7 @@ endif
 docs-image: docs-jdk
 
 # This target builds the test image
-test-image: prepare-test-image test-image-jdk-jtreg-native test-image-demos-jdk
+test-image: prepare-test-image test-image-jdk-jtreg-native test-image-demos-jdk test-image-libtest-jtreg-native
 
 ifneq ($(JVM_TEST_IMAGE_TARGETS), )
   # If JVM_TEST_IMAGE_TARGETS is externally defined, use it instead of the
