+++ test/hotspot/jtreg/testlibrary_tests/jvmti/SimpleClassFileLoadHookTest.java
