+++ test/lib-test/RedefineClassTest.java
