+++ test/lib-test/TEST.groups
@@ -21,18 +21,4 @@
 # questions.
 #
 
-# This file identifies the root of the test-suite hierarchy.
-# It also contains test-suite configuration information.
-
-# The list of keywords supported in this test suite
-# randomness:           test uses randomness, test cases differ from run to run
-keys=randomness
-
-# Minimum jtreg version
-requiredVersion=5.1 b1
-
-# Path to libraries in the topmost test directory. This is needed so @library
-# does not need ../../ notation to reach them
-external.lib.roots = ../../
-
-groups=TEST.groups
+tier1 = .
