+++ test/lib-test/jdk/test/lib/AssertsTest.java
