+++ test/lib-test/jdk/test/lib/TestPlatformIsTieredSupported.java
