+++ test/lib-test/sun/hotspot/whitebox/BlobSanityTest.java
