+++ test/lib-test/sun/hotspot/whitebox/CPUInfoTest.java
