+++ test/lib-test/sun/hotspot/whitebox/vm_flags/DoubleTest.java
