+++ test/lib-test/sun/hotspot/whitebox/vm_flags/VmFlagTest.java
